Removing pointless browser utils
