--- src/shared/utils/browser/can-share.ts
@@ -1,5 +1,0 @@
-import { isBrowser } from "@utils/browser";
-
-export default function canShare() {
-  return isBrowser() && !!navigator.canShare;
-}
