--- src/shared/utils/browser/clear-auth-cookie.ts
@@ -1,10 +1,0 @@
-import * as cookie from "cookie";
-import { authCookieName } from "../../config";
-
-export default function clearAuthCookie() {
-  document.cookie = cookie.serialize(authCookieName, "", {
-    maxAge: -1,
-    sameSite: "lax",
-    path: "/",
-  });
-}
