+++ src/shared/utils/browser/index.ts
@@ -1,27 +1,133 @@
-import canShare from "./can-share";
-import clearAuthCookie from "./clear-auth-cookie";
-import dataBsTheme from "./data-bs-theme";
-import isBrowser from "./is-browser";
-import isDark from "./is-dark";
-import nextUserAction from "./next-user-action";
-import platform from "./platform";
-import refreshTheme from "./refresh-theme";
-import setAuthCookie from "./set-auth-cookie";
-import setThemeOverride from "./set-theme-override";
-import share from "./share";
-import snapToTop from "./snap-to-top";
+import * as cookie from "cookie";
+import { authCookieName } from "../../config";
+import { GetSiteResponse, MyUserInfo } from "lemmy-js-client";
+import { isHttps } from "@utils/env";
 
-export {
-  canShare,
-  clearAuthCookie,
-  dataBsTheme,
-  isBrowser,
-  isDark,
-  nextUserAction,
-  platform,
-  refreshTheme,
-  setAuthCookie,
-  setThemeOverride,
-  share,
-  snapToTop,
-};
+export function canShare() {
+  return isBrowser() && !!navigator.canShare;
+}
+
+export function clearAuthCookie() {
+  document.cookie = cookie.serialize(authCookieName, "", {
+    maxAge: -1,
+    sameSite: "lax",
+    path: "/",
+  });
+}
+
+export function dataBsTheme(
+  siteResOrTheme?: GetSiteResponse | string,
+  myUserInfo?: MyUserInfo,
+) {
+  const theme =
+    typeof siteResOrTheme === "string"
+      ? siteResOrTheme
+      : (myUserInfo?.local_user_view.local_user.theme ??
+        siteResOrTheme?.site_view.local_site.default_theme ??
+        "browser");
+
+  return (isDark() && theme === "browser") || theme.includes("dark")
+    ? "dark"
+    : "light";
+}
+
+export function isBrowser() {
+  return typeof window !== "undefined";
+}
+
+export function isDark() {
+  return (
+    isBrowser() && window.matchMedia("(prefers-color-scheme: dark)").matches
+  );
+}
+
+const eventTypes = ["mousedown", "keydown", "touchstart", "touchmove", "wheel"];
+
+const scrollThreshold = 2;
+
+type Continue = boolean | void;
+
+export function nextUserAction(cb: (e: Event) => Continue) {
+  const eventTarget = window.document.body;
+
+  let cleanup: (() => void) | undefined = () => {
+    cleanup = undefined;
+    eventTypes.forEach(ev => {
+      eventTarget.removeEventListener(ev, listener);
+    });
+    window.removeEventListener("scroll", scrollListener);
+  };
+
+  const listener = (e: Event) => {
+    if (!cb(e)) {
+      cleanup?.();
+    }
+  };
+  eventTypes.forEach(ev => {
+    eventTarget.addEventListener(ev, listener);
+  });
+
+  let remaining = scrollThreshold;
+  const scrollListener = (e: Event) => {
+    // This only has to cover the scrollbars. The problem is that scroll events
+    // are also fired when the document height shrinks below the current bottom
+    // edge of the window.
+    remaining--;
+    if (remaining < 0) {
+      if (!cb(e)) {
+        cleanup?.();
+      } else {
+        remaining = scrollThreshold;
+      }
+    }
+  };
+  window.addEventListener("scroll", scrollListener);
+
+  return () => {
+    cleanup?.();
+  };
+}
+
+const platformString = () =>
+  navigator.platform?.match(/mac|win|linux/i)?.[0].toLowerCase();
+const getPlatformPredicate = (platform: string) => () =>
+  isBrowser() && platformString() === platform;
+const isWin = getPlatformPredicate("win");
+const isMac = getPlatformPredicate("mac");
+const isLinux = getPlatformPredicate("linux");
+
+export const platform = { isWin, isMac, isLinux };
+
+export function refreshTheme() {
+  if (isBrowser()) {
+    window.dispatchEvent(new CustomEvent("refresh-theme"));
+  }
+}
+
+export function setAuthCookie(jwt: string) {
+  document.cookie = cookie.serialize(authCookieName, jwt, {
+    maxAge: 365 * 24 * 60 * 60 * 1000,
+    secure: isHttps(),
+    sameSite: "lax",
+    path: "/",
+  });
+}
+
+export async function setThemeOverride(theme?: string) {
+  if (!isBrowser()) {
+    return;
+  }
+  window.dispatchEvent(
+    new CustomEvent("set-theme-override", { detail: { theme } }),
+  );
+}
+
+export function share(shareData: ShareData) {
+  if (isBrowser()) {
+    navigator.share(shareData);
+  }
+}
+
+export function snapToTop() {
+  window.scrollTo({ left: 0, top: 0, behavior: "instant" });
+}
