--- src/shared/utils/browser/is-browser.ts
@@ -1,3 +1,0 @@
-export default function isBrowser() {
-  return typeof window !== "undefined";
-}
