--- src/shared/utils/browser/is-dark.ts
@@ -1,7 +1,0 @@
-import isBrowser from "./is-browser";
-
-export default function isDark() {
-  return (
-    isBrowser() && window.matchMedia("(prefers-color-scheme: dark)").matches
-  );
-}
