--- src/shared/utils/browser/refresh-theme.tsx
@@ -1,7 +1,0 @@
-import { isBrowser } from ".";
-
-export default function refreshTheme() {
-  if (isBrowser()) {
-    window.dispatchEvent(new CustomEvent("refresh-theme"));
-  }
-}
