--- src/shared/utils/browser/share.ts
@@ -1,7 +1,0 @@
-import { isBrowser } from "@utils/browser";
-
-export default function share(shareData: ShareData) {
-  if (isBrowser()) {
-    navigator.share(shareData);
-  }
-}
