--- src/shared/utils/browser/snap-to-top.ts
@@ -1,3 +1,0 @@
-export default function snapToTop() {
-  window.scrollTo({ left: 0, top: 0, behavior: "instant" });
-}
